align middle arrows
